pyproject.toml
[tool.poetry]
name = "azure-hack-2024"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
ipykernel = "^6.29.4"
openai = "^1.33.0"
pillow = "^10.3.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
azure-search-documents = "^11.4.0"
flask = "^3.0.3"
langchain = "^0.2.3"
opencensus-ext-azure = "^1.1.13"
ruff = "^0.4.8"
pytest = "^8.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
